+++ Peer.cs
@@ -10,6 +10,7 @@ readonly struct Peer(ZNetPeer peer)
     public ZDOID m_characterID => _peer?.m_characterID ?? default;
     //public bool IsConnected => _peer?.m_socket.IsConnected() ?? default; // potentially takes a long time?
     public string GetHostName() => _peer?.m_socket.GetHostName() ?? "";
+    public IReadOnlyDictionary<string, string> m_serverSyncedPlayerData => _peer is { m_server: false } ? _peer.m_serverSyncedPlayerData : ZNet.instance.m_serverSyncedPlayerData;
 
     public override bool Equals(object obj) => Equals(_peer, obj);
     public override int GetHashCode() => _peer?.GetHashCode() ?? default;