+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v0.2.21
+- New feature: Allow unlocking of crypt doors if Elder has been defeated [#50](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/50)
 - Fix for [#53](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/53): Unable to use tools/weapons
 
 ### v0.2.20